# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CYFY News Board API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threats:
    get:
      operationId: getThreats
      tags: [threats]
      summary: List threat intelligence items
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [local, global]
        - name: severity
          in: query
          description: Comma-separated severities (e.g. critical,high,medium)
          schema:
            type: string
        - name: category
          in: query
          description: Comma-separated categories (e.g. ransomware,phishing)
          schema:
            type: string
        - name: state
          in: query
          description: Indian state code filter (e.g. MH, KA)
          schema:
            type: string
        - name: sector
          in: query
          description: Indian sector filter
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, monitoring]
        - name: timeframe
          in: query
          description: Time range shorthand (1h, 6h, 24h, 7d, 30d, all)
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d, all]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated threats list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatIntelListResponse"
  /threats/{id}:
    get:
      operationId: getThreatById
      tags: [threats]
      summary: Get threat by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Threat detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatIntelItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /threats/export:
    get:
      operationId: exportThreats
      tags: [threats]
      summary: Export threats as CSV
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [local, global]
        - name: severity
          in: query
          description: Comma-separated severities (e.g. critical,high)
          schema:
            type: string
        - name: category
          in: query
          description: Comma-separated categories
          schema:
            type: string
        - name: state
          in: query
          description: Indian state code filter
          schema:
            type: string
        - name: sector
          in: query
          description: Indian sector filter
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    ThreatIntelItem:
      type: object
      required:
        - id
        - title
        - summary
        - description
        - scope
        - severity
        - category
        - source
        - ttps
        - iocs
        - malwareFamilies
        - affectedSystems
        - mitigations
        - targetSectors
        - targetRegions
        - status
        - confidenceLevel
        - publishedAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        scope:
          type: string
          enum: [local, global]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        category:
          type: string
        threatActor:
          type: string
          nullable: true
        threatActorAliases:
          type: array
          items:
            type: string
        targetSectors:
          type: array
          items:
            type: string
        targetRegions:
          type: array
          items:
            type: string
        ttps:
          type: array
          items:
            type: string
        iocs:
          type: array
          items:
            type: string
        malwareFamilies:
          type: array
          items:
            type: string
        affectedSystems:
          type: array
          items:
            type: string
        mitigations:
          type: array
          items:
            type: string
        source:
          type: string
        sourceUrl:
          type: string
          nullable: true
        references:
          type: array
          items:
            type: string
        campaignName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, resolved, monitoring]
        confidenceLevel:
          type: string
          enum: [confirmed, high, medium, low]
        firstSeen:
          type: string
          nullable: true
        lastSeen:
          type: string
          nullable: true
        publishedAt:
          type: string
        updatedAt:
          type: string
        isIndiaRelated:
          type: boolean
          nullable: true
        indiaConfidence:
          type: integer
          nullable: true
        indianState:
          type: string
          nullable: true
        indianStateName:
          type: string
          nullable: true
        indianCity:
          type: string
          nullable: true
        indianSector:
          type: string
          nullable: true
    ThreatIntelListResponse:
      type: object
      required:
        - items
        - total
        - page
        - limit
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ThreatIntelItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...